[Refactor] Users and Jobs MVC

+++ backend/controllers/jobController.js
@@ -1,15 +1,19 @@
 const Job = require('../models/JobModel');
+const { JOB_STATUS } = require('../utils/constants');
 const BasicFilter = require('../utils/BasicFilter');
 const { handleAsync } = require('../utils/errorHandler');
 const AppError = require('../utils/AppError');
 
-exports.getJobs = handleAsync(async (req, res, next) => {
-	const filteredJob = new BasicFilter(Job.find(), req.query).filter().sort();
+exports.getAllActiveJobs = handleAsync(async (req, res, next) => {
+	const filter = { recruiter: req.user._id, status: JOB_STATUS.AVAILABLE };
+	const filteredJobs = new BasicFilter(Job.find(filter), req.query)
+		.filter()
+		.sort();
 
-	const job = await filteredJob.query;
-	res.status(201).json({
+	const jobs = await filteredJobs.query;
+	res.status(200).json({
 		status: 'success',
-		data: job,
+		data: jobs,
 	});
 });
 
@@ -23,12 +27,16 @@ exports.createJob = handleAsync(async (req, res, next) => {
 });
 
 exports.updateJob = handleAsync(async (req, res, next) => {
-	const job = await Job.findByIdAndUpdate(req.params.id, req.body, {
-		new: true,
-		runValidators: true,
-	});
+	let job = await Job.findById(req.params.id);
 	if (!job)
 		return next(new AppError('No such job exists. Update failed.', 404));
+	if (job.recruiter !== req.user._id)
+		return next(new AppError('Permission denied for this action', 403));
+
+	job = await Job.findByIdAndUpdate(req.params.id, req.body, {
+		runValidators: true,
+		new: true,
+	});
 	res.status(202).json({
 		status: 'success',
 		data: job,
@@ -36,16 +44,25 @@ exports.updateJob = handleAsync(async (req, res, next) => {
 });
 
 exports.deleteJob = handleAsync(async (req, res, next) => {
-	const job = await Job.findByIdAndDelete(req.params.id);
+	let job = await Job.findById(req.params.id);
 	if (!job)
 		return next(new AppError('No such job exists. Delete failed.', 404));
+	if (job.recruiter !== req.user._id)
+		return next(new AppError('Permission denied for this action', 403));
+
+	job = await Job.findByIdAndDelete(req.params.id);
+
 	res.status(204).json({
 		status: 'success',
 	});
 });
 
+exports.getJobs = handleAsync(async (req, res, next) => {
+	const filteredJob = new BasicFilter(Job.find(), req.query).filter().sort();
 
-exports.getAllActiveJobs = handleAsync(async(req, res, next) => {
-	
-
-})
+	const job = await filteredJob.query;
+	res.status(201).json({
+		status: 'success',
+		data: job,
+	});
+});
